+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* fix: Checks dialog wait for confirmation before upload
 
 0.3.0 (2022-05-10)
 ==================